// src/afreecatv/types/replay.ts
export type AfreecaTVReplay = { data: Replay[] };

type Replay = {
  title_no: number;
  station_no: string;
  station_user_id: string;
  bbs_no: string;
  board_type: string;
  auth_no: number;
  title_name: string;
  user_id: string;
  user_nick: string;
  profile_image: string;
  photo_cnt: string;
  platform_type: string;
  notice_yn: string;
  comment_yn: string;
  livepost_yn: string;
  search_yn: string;
  ip_inet: string;
  reg_date: string;
  reg_timestamp: null;
  content: string;
  like_cnt: string;
  read_cnt: string;
  comment_cnt: string;
  ucc_favorite_cnt: string;
  copyright: null;
  thumb: string;
  pc_thumb: string;
  scheme: string;
  flag: string;
  ucc_type: string;
  category: string;
  grade: string;
  is_adult: boolean;
  last_view_duration: string;
  duration: string;
  total_duration: number;
  file_type: string;
  is_fanclub: boolean;
  is_subscribe: boolean;
  is_ppv: boolean;
  url: string;
  web_path: string | null;
};